Make -t:doo support running too
