+++ Source/DafnyCore/Compilers/ExecutableBackend.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
@@ -111,5 +112,64 @@ public abstract class ExecutableBackend : Plugins.IExecutableBackend {
   protected static void WriteFromFile(string inputFilename, TextWriter outputWriter) {
     var rd = new StreamReader(new FileStream(inputFilename, FileMode.Open, FileAccess.Read));
     SinglePassCompiler.WriteFromStream(rd, outputWriter);
+  }
+
+  protected bool RunTargetDafnyProgram(string targetFilename, TextWriter outputWriter) {
+    
+    /*
+     * In order to work for the continuous integration, we need to call the Dafny compiler using dotnet
+     * because dafny is not currently in the path
+     */
+
+    var where = System.Reflection.Assembly.GetExecutingAssembly().Location;
+    where = System.IO.Path.GetDirectoryName(where);
+    var dafny = where + "/Dafny.dll";
+
+    var opt = Options;
+    var psi = PrepareProcessStartInfo("dotnet", opt.MainArgs
+      .Prepend("/compileTarget:cs")
+      .Prepend("/compile:4")
+      .Prepend("/compileVerbose:0")
+      .Prepend("/printVerifiedProceduresCount:0")
+      .Prepend("/noVerify")
+      .Prepend(targetFilename)
+      .Prepend(dafny));
+
+    /*
+     * When this code was written, the Dafny compiler cannot be made completely silent.
+     * This is a problem for this specific compiler and the integration tests because the second
+     * call to the compiler makes unexpected writes to the output.
+     * The following code is catching the output from the second compiler call (the one that executes the code)
+     * and stripping out the first two lines and the last line.
+     */
+
+    psi.RedirectStandardOutput = true;
+    var process = new Process();
+    process.StartInfo = psi;
+    var outputBuilder = new List<string>();
+    process.OutputDataReceived += (sender, args) => outputBuilder.Add(args.Data);
+
+    try {
+      process.Start();
+      process.BeginOutputReadLine();
+      process.WaitForExit();
+      process.CancelOutputRead();
+
+      for (int i = 2; i < outputBuilder.Count - 1; i++) {
+        Console.WriteLine(outputBuilder[i]);
+      }
+
+      if (process.ExitCode != 0) {
+        outputWriter.WriteLine("{0} Process exited with exit code {1}", "", process.ExitCode);
+        return false;
+      }
+
+    } catch (System.ComponentModel.Win32Exception e) {
+      string additionalInfo = $": {e.Message}";
+      outputWriter.WriteLine($"Error: Unable to start {psi.FileName}{additionalInfo}");
+      return false;
+    }
+
+    return true;
   }
 }
